requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
requests==2.32.5
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.9
ultralytics==8.3.70
pillow==11.0.0
huggingface-hub==0.26.2
vllm==0.6.4
git+https://github.com/ChaoningZhang/MobileSAM.git
